ptb updates
