+++ upgrade_policy/quorum_upgrade/scripts/ts/src/transactions.ts
@@ -1,5 +1,5 @@
 
-import { getUpgradeDigest, getActiveAddress, prepareAddressVecSet, signAndExecute } from "./utils";
+import { getUpgradeDigest, getActiveAddress, prepareAddressVecSet, prepareMetadataVecMap, signAndExecute } from "./utils";
 import { TransactionArgument, TransactionBlock } from '@mysten/sui.js/transactions';
 
 // =================================================================
@@ -7,28 +7,23 @@ import { TransactionArgument, TransactionBlock } from '@mysten/sui.js/transactio
 // =================================================================
 
 // Voters. Add all addresses that will be part of the quorum policy.
-const VOTER_1 = '0x754a43970ec22d78110069ca1b45fb20818348363648b6b486edf0045dea559e';
-const VOTER_2 = '0xe55684e6a538d29df4b9f510a6bd3ec17d3dc47fb524839954bf7fcf835b5d0d';
-const VOTER_3 = '0xe0b97bff42fcef320b5f148db69033b9f689555348b2e90f1da72b0644fa37d0';
+const VOTER_1 = '';
+const VOTER_2 = '';
+const VOTER_3 = '';
 // The package id of the `quorum_upgrade_policy` package (published on mainnet)
-const QUORUM_UPGRADE_PACKAGE_ID = `0x03e8badab460af8b3d4c96b283c4c3994b370c9dd2c1a4e5f80d649c25d305da`;
+const QUORUM_UPGRADE_PACKAGE_ID = ``;
+// Optional upgrade metadata info to be included with upgrade
 const UPGRADE_METADATA: Map<string, Uint8Array> = new Map();
-// Example metadata to be attached
-const text = 'Hello, World!';
-const encoder = new TextEncoder();
-const uint8FromString = encoder.encode(text);
-UPGRADE_METADATA.set('key1', uint8FromString);
-UPGRADE_METADATA.set('key2', new Uint8Array([10, 11, 12]));
 // The upgrade cap of the quorum upgrade policy (resulting from `quorum_upgrade_policy::new`)
-const QUORUM_UPGRADE_CAP_ID = `0x8d45558b0cdd2a0590d4c148817998aeb6b8f46fd9865436a1d20c30595c7a8f`;
+const QUORUM_UPGRADE_CAP_ID = ``;
 // path to the package to publish or upgrade
-const PATH_TO_PACKAGE = '/Users/tonylee/Documents/apps/upgrade_policy/quorum_upgrade';
+const PATH_TO_PACKAGE = '';
 // The package id of the testing package to be upgraded. That is the package id of the package
 // defined at the path above
 const TEST_PACKAGE_ID = ``;
 // The upgrade cap of the package to protect with the quorum upgrade policy
 // That is the upgrade cap of the package defined at the path above
-const TEST_PACKAGE_UPGRADE_CAP_ID = `0xf643c9af74a77e4435d5ed3e728756f422038844143f96c665fb9e2d4b3dfa60`;
+const TEST_PACKAGE_UPGRADE_CAP_ID = ``;
 // Voting cap used for voting.
 // That is the voting cap of the transaction signer
 const VOTING_CAP_ID = ``;
@@ -74,7 +69,9 @@ const proposeUpgrade = (txb: TransactionBlock, quorumUpgradeCapId: string, packa
 
 }
 
-const proposeUpgradeV2 = (txb: TransactionBlock, quorumUpgradeCapId: string, packagePath: string, metadata: Map<string, Uint8Array>) => {
+/// Calls `quorum_upgrade_policy::propose_upgrade_v2`, `quorum_upgrade_policy::add_all_metadata` (optional), `quorum_upgrade_policy::share_upgrade_object`,  
+/// Calling this will publish the `ProposedUpgrade` shared object.
+const proposeUpgradeV2 = (txb: TransactionBlock, quorumUpgradeCapId: string, packagePath: string, metadata = new Map<string, Uint8Array>) => {
     const { digest }  = getUpgradeDigest(packagePath);
 
     const proposedUpgrade = txb.moveCall({
@@ -84,16 +81,16 @@ const proposeUpgradeV2 = (txb: TransactionBlock, quorumUpgradeCapId: string, pac
             txb.pure(digest, 'vector<u8>')
         ]
     });
-
-    txb.moveCall({
-        target: `${QUORUM_UPGRADE_PACKAGE_ID}::quorum_upgrade_policy::add_all_metadata`,
-        arguments: [
-            txb.object(proposedUpgrade),
-            prepareMetadataVecMap(txb, metadata)
-        ]
-    });
-
-    console.log("Proposed Upgrade Object:", proposedUpgrade);
+    
+    if (metadata.size > 0) {
+        txb.moveCall({
+            target: `${QUORUM_UPGRADE_PACKAGE_ID}::quorum_upgrade_policy::add_all_metadata`,
+            arguments: [
+                txb.object(proposedUpgrade),
+                prepareMetadataVecMap(txb, metadata)
+            ]
+        });
+    }
 
     txb.moveCall({
         target: `${QUORUM_UPGRADE_PACKAGE_ID}::quorum_upgrade_policy::share_upgrade_object`,
@@ -102,27 +99,6 @@ const proposeUpgradeV2 = (txb: TransactionBlock, quorumUpgradeCapId: string, pac
         ]
     });
 }
-
-const prepareMetadataVecMap = (txb: TransactionBlock, metadata: Map<string, Uint8Array>): TransactionArgument => {
-    const vecMap = txb.moveCall({
-        target: `0x2::vec_map::empty<string::String, vector<u8>>`,
-        typeArguments: ['0x2::sui::vec_map']
-    });
-
-    metadata.forEach((value, key) => {
-        txb.moveCall({
-            target: `0x2::vec_map::insert`,
-            arguments: [
-                vecMap,
-                txb.pure.string(key),
-                txb.pure(value, 'vector<u8>'),
-            ],
-            typeArguments: ['0x1::string::String', 'vector<u8>']
-        });
-    });
-    return vecMap;
-}
-
 
 
 /// Vote for a particular `ProposedUpgrade` shared object.
@@ -137,7 +113,6 @@ const vote = (txb: TransactionBlock, proposedUpgradeObjectId: string, votingCapO
         ]
     });
 }
-
 
 /// Executes a `package upgrade`.
 /// It fails if the `ProposedUpgrade` object has not reached quorum.
@@ -169,26 +144,25 @@ const authorizeUpgrade = (txb: TransactionBlock, packageId: string, proposedUpgr
     });
 }
 
-
 /// Main entry points, comment out as needed...
 const executeTransaction = async () => {
 
     const txb = new TransactionBlock();
 
-    // // 1- define a 2 out of 3 quorum upgrade policy
-    // newQuorumUpgradeCap(txb, 2, [VOTER_1, VOTER_2, VOTER_3], TEST_PACKAGE_UPGRADE_CAP_ID, getActiveAddress());
+    // 1- define a 2 out of 3 quorum upgrade policy
+    newQuorumUpgradeCap(txb, 2, [VOTER_1, VOTER_2, VOTER_3], TEST_PACKAGE_UPGRADE_CAP_ID, getActiveAddress());
 
     // 2- propose an upgrade. Digest is determined automatically via the package path
     proposeUpgradeV2(txb, QUORUM_UPGRADE_CAP_ID, PATH_TO_PACKAGE, UPGRADE_METADATA);
+    
+    // 3- vote for an upgrade
+    vote(txb, PROPOSED_UPGRADE_ID, VOTING_CAP_ID);
 
-    // // 3- vote for an upgrade
-    // vote(txb, PROPOSED_UPGRADE_ID, VOTING_CAP_ID);
-
-    // // 4- authorize/commit the upgrade
-    // authorizeUpgrade(txb, TEST_PACKAGE_ID, PROPOSED_UPGRADE_ID, QUORUM_UPGRADE_CAP_ID, PATH_TO_PACKAGE);
+    // 4- authorize/commit the upgrade
+    authorizeUpgrade(txb, TEST_PACKAGE_ID, PROPOSED_UPGRADE_ID, QUORUM_UPGRADE_CAP_ID, PATH_TO_PACKAGE);
 
     // Run against mainnet
-    const res = await signAndExecute(txb, 'testnet');
+    const res = await signAndExecute(txb, 'mainnet');
 
     console.dir(res, { depth: null });
 }
