+++ upgrade_policy/quorum_upgrade/scripts/ts/src/utils.ts
@@ -79,6 +79,27 @@ export const prepareAddressVecSet = (txb: TransactionBlock, voters: string[]): T
     return vecSet;
 }
 
+/// Construct a VecMap of (string, vector<u8>) key-value pairs.
+export const prepareMetadataVecMap = (txb: TransactionBlock, metadata: Map<string, Uint8Array>): TransactionArgument => {
+    const vecMap = txb.moveCall({
+        target: `0x2::vec_map::empty`,
+        typeArguments: ['0x1::string::String', 'vector<u8>']
+    });
+
+    metadata.forEach((value, key) => {
+        txb.moveCall({
+            target: `0x2::vec_map::insert`,
+            arguments: [
+                vecMap,
+                txb.pure.string(key),
+                txb.pure(Array.from(value), 'vector<u8>'),
+            ],
+            typeArguments: ['0x1::string::String', 'vector<u8>']
+        });
+    });
+    return vecMap;
+}
+
 /// A helper to sign & execute a transaction.
 export const signAndExecute = async (txb: TransactionBlock, network: Network) => {
     const client = getClient(network);
